LayoutTests/inspector/styles/parse-utf8-bom.html

<html>
<head>
<link rel="stylesheet" type="text/css" href="resources/parse-utf8-bom-main.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.selectNodeAndWaitForStyles("inspected", step1);

    function step1()
    {
        InspectorTest.dumpSelectedElementStyles(true, false, true);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that source data are extracted correctly from external stylesheets in UTF-8 with BOM. <a href="https://bugs.webkit.org/show_bug.cgi?id=59322">Bug 59322</a>.
</p>
<h1 id="inspected">
I'm red.
</h1>

</body>
</html>
