LayoutTests/inspector/styles/parse-utf8-bom-expected.txt

Tests that source data are extracted correctly from external stylesheets in UTF-8 with BOM. Bug 59322.

I'm red.

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
h1  { (parse-utf8-bom-main.css:3)
color: red;

[expanded] 
h1  { (user agent stylesheet)
display: block;
font-size: 2em;
-webkit-margin-before: 0.67em;
-webkit-margin-after: 0.67em;
-webkit-margin-start: 0px;
-webkit-margin-end: 0px;
font-weight: bold;
